requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.8.1
pydantic==2.8.2
